Add warehouses tool to @dust
